move status_helper out of cmds dir (#186)
